Zoom2 Fix serial gpmc setup
The offset to the chip select is incorrect.

The change 5c691694eabf90db6d6d11fd8be65c296b7324f9,

omap3: embedd gpmc_cs into gpmc config struct

introduced a problem with the serial gpmc setup.

This patch reverts the chip select to its previous value.

The symptoms of this problem are that the Zoom2
currently hangs.

This was run tested on Zoom2.

Signed-off-by: Tom Rix <Tom.Rix@windriver.com>